+++ .vscode/settings.json
@@ -1,0 +1,5 @@
+{
+  "path-intellisense.mappings": {
+    "@": "${workspaceFolder}/src"
+  }
+}
