+++ manifest.json
@@ -2,6 +2,7 @@
   "name": "Team Library Component Browser",
   "id": "813970431341620710",
   "api": "1.0.0",
+  "editorType": ["figma"],
   "main": "dist/code.js",
   "ui": "dist/ui.html"
 }
