+++ package.json
@@ -19,6 +19,7 @@
     "@types/react-modal": "^3.10.5",
     "@typescript-eslint/eslint-plugin": "^2.20.0",
     "@typescript-eslint/parser": "^2.20.0",
+    "clean-webpack-plugin": "^4.0.0",
     "css-loader": "^3.4.2",
     "eslint": "^6.8.0",
     "eslint-config-prettier": "^6.10.0",
