+++ src/ui/components/Search.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react'
 import { inject, observer } from 'mobx-react'
-import Store from '@/app/Store'
+import Store from '@/ui/Store'
 import _ from 'lodash'
 
 type Props = {
@@ -92,7 +92,7 @@ export default class Search extends React.Component<Props, State> {
       <div className="search">
         <img
           className="search-icon"
-          src={require('@/app/assets/img/icon_search.svg').default}
+          src={require('@/ui/assets/img/icon_search.svg').default}
           alt=""
         />
         <input
@@ -109,10 +109,7 @@ export default class Search extends React.Component<Props, State> {
           }`}
           onClick={this.onClearClick.bind(this)}
         >
-          <img
-            src={require('@/app/assets/img/icon_close.svg').default}
-            alt=""
-          />
+          <img src={require('@/ui/assets/img/icon_close.svg').default} alt="" />
         </div>
       </div>
     )
