+++ src/ui/main.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react'
 import * as ReactDOM from 'react-dom'
-import App from '@/app/components/App'
+import App from '@/ui/components/App'
 import Modal from 'react-modal'
 
 const app = document.getElementById('app') as HTMLElement
