+++ tsconfig.json
@@ -15,7 +15,6 @@
     "noImplicitAny": false,
     "baseUrl": ".",
     "paths": {
-      "~/*": ["./src/*"],
       "@/*": ["./src/*"],
       "*": ["./@types/*"]
     },
