+++ webpack.config.js
@@ -3,20 +3,24 @@ const path = require('path')
 const HtmlWebpackInlineSourcePlugin = require('html-webpack-inline-source-plugin')
 const HtmlWebpackPlugin = require('html-webpack-plugin')
 const TerserPlugin = require('terser-webpack-plugin')
+const { CleanWebpackPlugin } = require('clean-webpack-plugin')
 
 module.exports = (env, argv) => ({
   mode: argv.mode === 'production' ? 'production' : 'development',
   devtool: argv.mode === 'production' ? false : 'inline-source-map',
   entry: {
-    ui: './src/app/index.tsx', // The entry point for your UI code
-    code: './src/plugin/controller.ts' // The entry point for your plugin code
+    code: './src/code.ts', // The entry point for your plugin code
+    ui: './src/ui/main.tsx' // The entry point for your UI code
+  },
+  output: {
+    path: path.resolve(__dirname, 'dist'),
+    filename: '[name].js'
   },
   module: {
     rules: [
       {
         test: /\.tsx?$/,
-        loader: 'ts-loader',
-        exclude: /node_modules/
+        loader: 'ts-loader'
       },
       {
         test: /\.css$/,
@@ -50,10 +54,19 @@ module.exports = (env, argv) => ({
   resolve: {
     extensions: ['.tsx', '.ts', '.jsx', '.js'],
     alias: {
-      '@': path.resolve(__dirname, 'src'),
-      '~': path.resolve(__dirname, 'src')
+      '@': path.resolve(__dirname, 'src')
     }
   },
+  plugins: [
+    new CleanWebpackPlugin(),
+    new HtmlWebpackPlugin({
+      template: './src/ui/index.html',
+      filename: 'ui.html',
+      inlineSource: '.(js)$',
+      chunks: ['ui']
+    }),
+    new HtmlWebpackInlineSourcePlugin()
+  ],
   optimization: {
     minimizer: [
       new TerserPlugin({
@@ -70,18 +83,5 @@ module.exports = (env, argv) => ({
         }
       })
     ]
-  },
-  output: {
-    filename: '[name].js',
-    path: path.resolve(__dirname, 'dist')
-  },
-  plugins: [
-    new HtmlWebpackPlugin({
-      template: './src/app/index.html',
-      filename: 'ui.html',
-      inlineSource: '.(js)$',
-      chunks: ['ui']
-    }),
-    new HtmlWebpackInlineSourcePlugin()
-  ]
+  }
 })
